un tas de truc en plus

+++ include/libft.h
@@ -6,7 +6,7 @@
 /*   By: gajanvie <gajanvie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/13 18:28:39 by gajanvie          #+#    #+#             */
-/*   Updated: 2025/11/14 10:35:25 by gajanvie         ###   ########.fr       */
+/*   Updated: 2025/11/18 14:31:16 by gajanvie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef struct s_list
 	struct s_list	*next;
 }				t_list;
 
-unsigned long	ft_atol_base(char *str, char *base);
+unsigned int	ft_atol_base(char *str, char *base);
 int				ft_atoi(const char *nptr);
 void			ft_bzero(void *s, size_t n);
 void			*ft_calloc(size_t nmemb, size_t size);
